feat(codex): default-enable image generation bridge and support image edits
