+++ internal/management/authfiles/entry_test.go
@@ -193,8 +193,31 @@ func TestBuildEntryIncludesCodexOAuthAdmissionPayload(t *testing.T) {
 	if !ok {
 		t.Fatalf("codex_image_generation_bridge = %#v, want map", entry["codex_image_generation_bridge"])
 	}
+	// Absent metadata reports enabled: the executor injects the hosted tool unless an
+	// operator opts out, and the panel toggle has to show that same state.
+	if enabled, _ := bridge["enabled"].(bool); !enabled {
+		t.Fatalf("bridge.enabled = %#v, want true by default", bridge["enabled"])
+	}
+}
+
+func TestBuildEntryReportsCodexImageGenerationBridgeOptOut(t *testing.T) {
+	auth := &coreauth.Auth{
+		ID:       "codex-oauth",
+		Provider: "codex",
+		Attributes: map[string]string{
+			"runtime_only": "true",
+		},
+		Metadata: map[string]any{
+			"codex_image_generation_bridge": false,
+		},
+	}
+	entry := BuildEntry(auth, EntryOptions{})
+	bridge, ok := entry["codex_image_generation_bridge"].(map[string]any)
+	if !ok {
+		t.Fatalf("codex_image_generation_bridge = %#v, want map", entry["codex_image_generation_bridge"])
+	}
 	if enabled, _ := bridge["enabled"].(bool); enabled {
-		t.Fatalf("bridge.enabled = %#v, want false by default", bridge["enabled"])
+		t.Fatalf("bridge.enabled = %#v, want false when explicitly disabled", bridge["enabled"])
 	}
 }
